Make float_tolerance available
